% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_corpus.R
\name{calculate_doc_freq}
\alias{calculate_doc_freq}
\title{Calclulate document frequency}
\usage{
calculate_doc_freq(df, doc, term)
}
\arguments{
\item{df}{tidy data frame with one term per row}

\item{doc}{column with document id}

\item{term}{column with terms}
}
\value{
tidy data frame with added column df giving document frequencies for
        terms and df_ratio giving relative document frequency
}
\description{
Adds document frequencies to new column. Document frequency describes in how
many documents term appears. Useful in finding very common terms appearing
in almost all the documents and very rare terms appearing in only single or
very few documents.
}
