% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{compute_error}
\alias{compute_error}
\title{Title}
\usage{
compute_error(xgb_model, xgb_data_test)
}
\arguments{
\item{xgb_model}{model returned by xgb.train}

\item{xgb_data_test}{xgb.DMatrix test data}
}
\value{
mean error value
}
\description{
Title
}
