% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_to_dtm.R
\name{corpus_to_dtm}
\alias{corpus_to_dtm}
\title{Title}
\usage{
corpus_to_dtm(df, doc, term)
}
\arguments{
\item{df}{tidy data frame with one term per row}

\item{doc}{document column}

\item{term}{term column}
}
\value{
quanteda dfm object
}
\description{
Title
}
