% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtm.R
\docType{data}
\name{dtm}
\alias{dtm}
\alias{tc}
\alias{tf}
\alias{tf_idf}
\title{Document term matrices of aspol_filtered data set}
\format{
An object of class \code{dfm} with 61 rows and 4859 columns.

An object of class \code{dfm} with 61 rows and 4859 columns.

An object of class \code{dfm} with 61 rows and 4859 columns.
}
\usage{
tc

tf

tf_idf
}
\description{
Quanteda package defined format
}
\keyword{datasets}
