% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_corpus.R
\name{filter_upostag}
\alias{filter_upostag}
\title{Title}
\usage{
filter_upostag(df, upostag)
}
\arguments{
\item{df}{data frame}

\item{upostag}{character vector of upostag classes}
}
\value{
data frame
}
\description{
Title
}
