% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda.R
\name{get_coherence}
\alias{get_coherence}
\title{Title}
\usage{
get_coherence(dfm, k)
}
\arguments{
\item{dfm}{quanteda dfm object}

\item{k}{number of topics to classify}
}
\value{
named vector of k and coherence score
}
\description{
Title
}
