% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda.R
\name{get_doc_topic_prob}
\alias{get_doc_topic_prob}
\title{Title}
\usage{
get_doc_topic_prob(dfm, k)
}
\arguments{
\item{dfm}{quanteda dfm object}

\item{k}{number of topics to classify}
}
\value{
data frame of document topic probablilities
}
\description{
Title
}
