% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{get_train_test_data}
\alias{get_train_test_data}
\title{Title}
\usage{
get_train_test_data(x, y, trainNames)
}
\arguments{
\item{x}{quanteda dfm object}

\item{y}{data frame with classes}

\item{trainNames}{character vector of document names}
}
\value{
list of xgb.DMatrices for training and testing
}
\description{
Title
}
