% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kuntalista.R
\docType{data}
\name{kuntalista}
\alias{kuntalista}
\title{Housing policy documents in municipalities}
\format{
A data frame with rows and columns:
\describe{
  \item{kunta}{Kunta}
  \item{aspol}{Asuntopoliittinen ohjelma: Ohjelma laadittu, Ei ohjelmaa, Ei vastausta, Seutukuntaohjelma}
}
}
\usage{
kuntalista
}
\description{
Manually made list of municipalities and their housing policy documents
}
\examples{
kuntalista

}
\keyword{datasets}
