% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lda_models}
\alias{lda_models}
\title{Pre calculated semantic coherence scores for aspol data set}
\format{
A data frame with rows and columns:
\describe{
  \item{K}{Number of topics}
  \item{S}{Random seed}
  \item{mean_coherence}{Mean semantic topic coherence score}
}
}
\usage{
lda_models
}
\description{
Mean semantic topic coherence score for different combinations of k and
random seed.
}
\examples{
lda_models


}
\keyword{datasets}
