% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{mod}
\alias{mod}
\title{Title}
\usage{
mod(xgb_data_train, ...)
}
\arguments{
\item{xgb_data_train}{xgb.DMatrix with train data}

\item{...}{xgb model parameters}
}
\value{
xgb model
}
\description{
Title
}
