% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_corpus.R
\name{preprocess_corpus}
\alias{preprocess_corpus}
\title{Preprocess corpus object}
\usage{
preprocess_corpus(df, doc, term)
}
\arguments{
\item{df}{Data frame in tidytext format one token per document per line}

\item{doc}{column with document id}

\item{term}{column with terms}
}
\value{
cleaned data frame
}
\description{
Preprocessing pipeline:
1. Filter short terms
2. Filter common terms appearing in most documents
3. Filter terms appearing in few documents
4. Filter only nouns
}
\examples{
preprocess_corpus(aspol, kunta, LEMMA)

}
