% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_corpus.R
\name{remove_common_terms}
\alias{remove_common_terms}
\title{Title}
\usage{
remove_common_terms(df, doc = .data$kunta, ratio = 0.99)
}
\arguments{
\item{df}{data frame}

\item{doc}{document}

\item{ratio}{ratio for common terms, default 0.99}
}
\value{
data frame
}
\description{
Title
}
