% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_corpus.R
\name{remove_foreign}
\alias{remove_foreign}
\title{Remove foreign terms}
\usage{
remove_foreign(df)
}
\arguments{
\item{df}{tidy data frame with one term per row}
}
\value{
data frame with foreign terms removed
}
\description{
Foreign terms are defined as having UPOSTAX=="X" and FEATS=="Foreign=Yes"
}
