% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_corpus.R
\name{remove_numbers}
\alias{remove_numbers}
\title{Remove terms containing numbers}
\usage{
remove_numbers(df, term)
}
\arguments{
\item{df}{data frame}

\item{term}{column with the terms}
}
\value{
data frame
}
\description{
Remove terms containing numbers
}
