% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_corpus.R
\name{remove_short_term}
\alias{remove_short_term}
\title{Remove short terms}
\usage{
remove_short_term(df, len)
}
\arguments{
\item{df}{data frame}

\item{len}{minimum length for terms/words to keep}
}
\value{
data frame
}
\description{
Remove short terms
}
