% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_corpus.R
\name{remove_uncommon_terms}
\alias{remove_uncommon_terms}
\title{Title}
\usage{
remove_uncommon_terms(df, doc = .data$kunta, ratio = 0.05)
}
\arguments{
\item{df}{data frame}

\item{doc}{document}

\item{ratio}{ratio for rare terms, default 0.05}
}
\value{
data frame
}
\description{
Title
}
