% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanalista.R
\docType{data}
\name{sanalista}
\alias{sanalista}
\title{Official Finnish word list}
\format{
A data frame with rows and columns:
\describe{
  \item{Hakusana}{Hakusanat on poimittu Kielitoimiston sanakirjasta sellaisinaan.}
  \item{Homonymia}{Homonymia tarkoittaa toisiinsa liittymättömien sanojen satunnaista samannäköisyyttä.}
  \item{Sanaluokka}{Sanaluokka}
  \item{Taivutustiedot}{Taivutustiedoissa olevat numerot viittaavat sanan taivutustyyppeihin ja kirjaimet astevaihtelutyyppeihin.}
}
}
\usage{
sanalista
}
\description{
A dataset containing list of Finnish words
}
\details{
@source Kielitoimiston sanakirja,
 \url{https://www.kotus.fi/aineistot/sana-aineistot/nykysuomen_sanalista},
 Lisenssi: CC BY 4.0
}
\examples{
sanalista

}
\keyword{datasets}
