% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{split_train_test}
\alias{split_train_test}
\title{Title}
\usage{
split_train_test(data, classes, id, ratio = 0.7)
}
\arguments{
\item{data}{data frame}

\item{classes}{column giving classes}

\item{id}{column identifying documents}

\item{ratio}{ratio for training data, defaults to 0.7}
}
\value{
character vector of id names
}
\description{
Title
}
