% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_semantic_coherence.R
\name{calculate_semantic_coherence}
\alias{calculate_semantic_coherence}
\title{Title}
\usage{
calculate_semantic_coherence(dtm, K_values, n_seeds, seed, verbose = FALSE)
}
\arguments{
\item{dtm}{document term matrix (quanteda::dfm or tm::tm)}

\item{K_values}{vector for k values}

\item{n_seeds}{number of different random initializations for every k value}

\item{seed}{if seeds should be reproducible}

\item{verbose}{print extra imformation ot processing time}
}
\value{
tibble with LDA models and semantic coherence
}
\description{
Title
}
\examples{
dtm <- aspol |>
preprocess_corpus(kunta) |>
 dplyr::count(kunta, LEMMA) |>
 tidytext::cast_dfm(kunta, LEMMA, n)

 calculate_semantic_coherence(dtm = dtm, K_values = seq(5, 10, by = 5),
                              n_seeds = 1, seed = 1234, verbose = TRUE)
}
