% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kuntalista.R
\docType{data}
\name{kuntalista}
\alias{kuntalista}
\title{Housing policy documents in municipalities}
\format{
A data frame with rows and columns:
\describe{
  \item{Kunta}{Kunta}
  \item{Asuntopoliittinen ohjelma}{Kyllä, Ei, Ei mukana, Yhteinen}
}
}
\usage{
kuntalista
}
\description{
A dataset containing list municipalities and housing policy documents
}
\examples{
kuntalista

}
\keyword{datasets}
